requirements.txt
requests
web3
eth_utils
rich
colorama
python-dotenv
